Refactored authentication to its own package
