+++ app/main.py
@@ -9,7 +9,8 @@ load_dotenv()
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-from .routers import tokens, users, websockets
+from app.auth import router as auth_routes
+from .routers import users, websockets
 from .database import create_db_and_tables
 from .models import create_fake_users, create_admin_user
 
@@ -34,7 +35,7 @@ api.add_middleware(
     allow_methods=["*"],
     allow_headers=["*"],
 )
-api.include_router(tokens.router)
+api.include_router(auth_routes.router)
 api.include_router(users.router)
 api.include_router(websockets.router)
